Revert "Diberry/0126 auth (#2)"
This reverts commit 24be325675cc0d440edc3b7ad24b6d23ee5e1e99.
