+++ .env
@@ -1,0 +1,4 @@
+ENVIRONMENT=development
+DATABASE_NAME=
+DATABASE_COLLECTION_NAME=
+DATABASE_URL=
