+++ .env.sample
@@ -1,4 +1,3 @@
 ENVIRONMENT=development
 DATABASE_NAME=
 DATABASE_COLLECTION_NAME=
-DATABASE_URL=
