+++ .prettierignore
@@ -1,0 +1,7 @@
+dist
+coverage
+node_modules
+.devcontainer
+.vscode
+package-lock.json
+.next
