--- client/.eslintcache
@@ -1,1 +1,0 @@
-[{"C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\index.js":"1","C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\App.js":"2","C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\reportWebVitals.js":"3","C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\api.js":"4","C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\Form.js":"5","C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\List.js":"6"},{"size":500,"mtime":1611585007445,"results":"7","hashOfConfig":"8"},{"size":2023,"mtime":1611683012487,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":1611585007445,"results":"10","hashOfConfig":"8"},{"size":1158,"mtime":1611682921857,"results":"11","hashOfConfig":"8"},{"size":1066,"mtime":1611683297563,"results":"12","hashOfConfig":"8"},{"size":715,"mtime":1611670051506,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"10ip7id",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\index.js",[],["27","28"],"C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\App.js",[],"C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\reportWebVitals.js",[],"C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\api.js",[],"C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\Form.js",[],"C:\\repos\\js-e2e\\js-e2e-express-mongodb\\client\\src\\List.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
