--- server/.eslintrc
@@ -1,41 +1,0 @@
-{
-  "plugins": [
-    "prettier"
-  ],
-  "root": true,
-  "extends": [
-    "eslint:recommended",
-    "airbnb-base",
-    "prettier"
-  ],
-  "env": {
-    "commonjs": true,
-    "es2020": true,
-    "node": true,
-    "jest": true
-  },
-  "parserOptions": {
-    "ecmaVersion": 11
-  },
-  "ignorePatterns": [
-    ".vscode",
-    ".devcontainer",
-    "dist"
-  ],
-  "rules": {
-    "semi": [
-      "warn",
-      "always"
-    ],
-    "quotes": [
-      "error",
-      "single"
-    ],
-    "no-console": [
-      0
-    ],
-    "no-return-await": [
-      "warn"
-    ]
-  }
-}
