--- server/.prettierignore
@@ -1,7 +1,0 @@
-dist
-coverage
-node_modules
-.devcontainer
-.vscode
-package-lock.json
-.next
