--- server/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "js-e2e-express-mongo",
-  "version": "1.0.0",
-  "description": "JavaScript server written with Express.js and the native MongoDB API",
-  "main": "index.js",
-  "directories": {
-    "doc": "docs",
-    "test": "test"
-  },
-  "scripts": {
-    "test": "npm run test:server && npm run test:client",
-    "test:server": "jest --detectOpenHandles --coverage",
-    "test:client": "cd ../client && npm run test",
-    "start": "nodemon src/index.js --unhandled-rejections=warn",
-    "start:server": "npm run build:server && npm run start",
-    "start:client": "cd ../client && npm start",
-    "start:full": "npm-run-all --parallel start:client start:server ",
-    "build:client": "cd ../client && npm run build",
-    "build:server": "eslint src/**/*.js --fix"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/diberry/js-e2e-express-mongo.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/diberry/js-e2e-express-mongo/issues"
-  },
-  "homepage": "https://github.com/diberry/js-e2e-express-mongo#readme",
-  "devDependencies": {
-    "eslint": "^7.9.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "jest": "^26.4.2",
-    "nodemon": "^2.0.4",
-    "prettier": "^2.1.2"
-  },
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "express-react-views": "^0.11.0",
-    "mongodb": "^3.6.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "serve-static": "^1.14.1"
-  },
-  "prettier": {
-    "printWidth": 80,
-    "semi": true,
-    "singleQuote": true,
-    "trailingComma": "es5"
-  }
-}
