--- server/src/index.js
@@ -1,12 +1,0 @@
-const server = require('./server');
-
-const port = process.env.PORT || 8080;
-
-server
-  .create()
-  .then((app) => {
-    app.listen(port, () => {
-      console.log(`Server has started on port ${port}!`);
-    });
-  })
-  .catch((err) => console.log(err));
